fix: rl mismatch reward type

+++ src/hpcsim/rl/gas_marl.py
@@ -196,7 +196,7 @@ class GASMARLAgent:
 
     def remember(self, state_t, value, lp1, lp2, ac1, ac2,
                  reward, m1, m2, job_input):
-        self._rewards.append(reward)
+        self._rewards.append(float(reward))
         self._states.append(state_t.to("cpu"))
         self._lp1.append(lp1.to("cpu"))
         self._lp2.append(lp2.to("cpu"))
@@ -219,7 +219,7 @@ class GASMARLAgent:
         return scipy.signal.lfilter([1], [1, -discount], x[::-1])[::-1]
 
     def finish_path(self, last_val=0.0):
-        rews   = np.append(np.array(self._rewards), last_val)
+        rews   = np.append(np.array(self._rewards, dtype=np.float32), float(last_val))
         values = torch.cat(self._values).squeeze(-1).numpy()
         vals   = np.append(values, last_val)
         deltas = rews[:-1] + self.gamma * vals[1:] - vals[:-1]
@@ -228,7 +228,7 @@ class GASMARLAgent:
         return adv.tolist(), ret.tolist()
 
     def commit_trajectory(self, last_reward: float):
-        adv, ret = self.finish_path(last_reward)
+        adv, ret = self.finish_path(float(last_reward))
         self.buffer.store(
             self._states, self._masks1, self._masks2,
             self._actions1, self._actions2,
