+++ src/hpcsim/rl/maskable_ppo.py
@@ -154,7 +154,7 @@ class MaskablePPOAgent:
         return int(logits.argmax(dim=-1).item())
 
     def remember(self, state_t, value, log_prob, action, reward, mask_t):
-        self._rewards.append(reward)
+        self._rewards.append(float(reward))
         self._states.append(state_t.to("cpu"))
         self._log_probs.append(log_prob.to("cpu"))
         self._values.append(value.to("cpu"))
@@ -172,7 +172,7 @@ class MaskablePPOAgent:
         return scipy.signal.lfilter([1], [1, -discount], x[::-1])[::-1]
 
     def finish_path(self, last_val: float = 0.0):
-        rews = np.append(np.array(self._rewards), last_val)
+        rews = np.append(np.array(self._rewards, dtype=np.float32), float(last_val))
         values = torch.cat(self._values).squeeze(-1).numpy()
         vals   = np.append(values, last_val)
         deltas = rews[:-1] + self.gamma * vals[1:] - vals[:-1]
